# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: metarang
    description: REST API for metarang microservices. Legacy-compatible JSON responses served via Kong → grpc-gateway → gRPC services.
    title: metarang API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: grpc-gateway direct (dev)
      url: http://localhost:8080
...
paths:
...
    /api/notifications/read/all:
        post:
            operationId: post_notifications_read_all
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Mark all notifications as read
            tags:
                - Notifications
...
    /api/search/users:
        get:
            operationId: get_search_users
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: Search users
            tags:
                - Search
        post:
            operationId: post_search_users
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: SearchUsers
            tags:
                - Search
...
    /api/user/profile:
        patch:
            operationId: patch_user_profile
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Update authenticated user profile
            tags:
                - Users
        put:
            operationId: put_user_profile
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Update authenticated user profile
            tags:
                - Users
...
    /api/user/wallet:
        get:
            operationId: get_user_wallet
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Get authenticated user wallet
            tags:
                - Users
    /api/users:
        get:
            operationId: get_users
            parameters:
                - description: order-by query parameter
                  in: query
                  name: order-by
                  required: false
                  schema:
                    type: string
                - description: Page number for pagination
                  in: query
                  name: page
                  required: false
                  schema:
                    type: integer
                - description: Search query
                  in: query
                  name: search
                  required: false
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List users
            tags:
                - Users
...
    /api/users/{user}/profile:
        get:
            operationId: get_users_user_profile
            parameters:
                - description: User ID or citizen code
                  in: path
                  name: user
                  required: true
                  schema:
                    example: user-code
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: GetUserProfile
            tags:
                - Users
...
    /api/users/{user}/wallet:
        get:
            operationId: get_users_user_wallet
            parameters:
                - description: User ID or citizen code
                  in: path
                  name: user
                  required: true
                  schema:
                    example: user-code
                    type: string
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: GetUserWallet
            tags:
                - Users
...
components:
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http